Separate ratchet history persistence
Keep the rolling-window model and statistics independent of persisted
JSON and filesystem details. Move schema validation, typed read errors,
and atomic history publication into a dedicated adapter without changing
the history artifact or ratchet CLI contracts.
