+++ tests/behaviour/test_benchmark_ratchet_noise_behaviour.py
@@ -17,8 +17,8 @@ from benchmarks.benchmark_profile import BENCHMARK_PROFILE_VERSION
 from benchmarks.ratchet_history import (
     BaselineHistory,
     HistorySample,
-    write_history,
 )
+from benchmarks.ratchet_history_persistence import write_history
 from benchmarks.ratchet_rust_performance import main as ratchet_cli
 from tests.behaviour._benchmark_ratchet_support import _write_json
